+++ src/components/Pokemon/Container.tsx
@@ -1,11 +1,11 @@
-import { Error, Loading, Render } from '.'
+import { Error, Loading, Pokemon } from '.'
 
 import { ContainerProps } from './types'
 
 export function Container({ isLoading, error, pokemon }: ContainerProps) {
   if (isLoading) return <Loading />
   if (error) return <Error />
-  if (pokemon) return <Render pokemon={pokemon} />
+  if (pokemon) return <Pokemon pokemon={pokemon} />
 
   return <></>
 }
